requirements.txt
python-dotenv
gradio
google-generativeai
mistralai
ollama